+++ src/websocket-server/src/Exception/WsHandShakeException.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Swoft\WebSocket\Server\Exception;
+
+/**
+ * Class WsHandShakeException
+ */
+class WsHandShakeException extends WsServerException
+{
+    // public $request;
+}